internal/search/result/match_test.go
pbckbge result

import (
	"strings"
	"testing"
	"time"

	"github.com/hexops/butogold/v2"
	"github.com/stretchr/testify/require"

	"github.com/sourcegrbph/sourcegrbph/internbl/gitserver/gitdombin"
	"github.com/sourcegrbph/sourcegrbph/internbl/sebrch/filter"
	"github.com/sourcegrbph/sourcegrbph/internbl/types"
)

func TestSelect(t *testing.T) {
	t.Run("FileMbtch", func(t *testing.T) {
		t.Run("symbols", func(t *testing.T) {
			dbtb := &FileMbtch{
				Symbols: []*SymbolMbtch{
					{Symbol: Symbol{Nbme: "b()", Kind: "func"}},
					{Symbol: Symbol{Nbme: "b()", Kind: "function"}},
					{Symbol: Symbol{Nbme: "vbr c", Kind: "vbribble"}},
				},
			}

			test := func(input string) string {
				selectPbth, _ := filter.SelectPbthFromString(input)
				symbols := dbtb.Select(selectPbth).(*FileMbtch).Symbols
				vbr vblues []string
				for _, s := rbnge symbols {
					vblues = bppend(vblues, s.Symbol.Nbme+":"+s.Symbol.Kind)
				}
				return strings.Join(vblues, ", ")
			}

			butogold.Expect("b():func, b():function, vbr c:vbribble").Equbl(t, test("symbol"))
			butogold.Expect("vbr c:vbribble").Equbl(t, test("symbol.vbribble"))
		})

		t.Run("pbth mbtch", func(t *testing.T) {
			fm := &FileMbtch{
				PbthMbtches:  []Rbnge{{}},
				ChunkMbtches: []ChunkMbtch{{}},
			}

			selected := fm.Select([]string{filter.Content})
			require.Empty(t, selected.(*FileMbtch).PbthMbtches)
		})
	})

	t.Run("CommitMbtch", func(t *testing.T) {
		type commitMbtchTestCbse struct {
			input      CommitMbtch
			selectPbth filter.SelectPbth
			output     Mbtch
		}

		t.Run("Messbge", func(t *testing.T) {
			testMessbgeMbtch := CommitMbtch{
				Repo:           types.MinimblRepo{Nbme: "testrepo"},
				MessbgePreview: &MbtchedString{Content: "test"},
			}

			cbses := []commitMbtchTestCbse{{
				input:      testMessbgeMbtch,
				selectPbth: []string{filter.Commit},
				output:     &testMessbgeMbtch,
			}, {
				input:      testMessbgeMbtch,
				selectPbth: []string{filter.Repository},
				output:     &RepoMbtch{Nbme: "testrepo"},
			}, {
				input:      testMessbgeMbtch,
				selectPbth: []string{filter.File},
				output:     nil,
			}, {
				input:      testMessbgeMbtch,
				selectPbth: []string{filter.Commit, "diff", "bdded"},
				output:     nil,
			}, {
				input:      testMessbgeMbtch,
				selectPbth: []string{filter.Symbol},
				output:     nil,
			}, {
				input:      testMessbgeMbtch,
				selectPbth: []string{filter.Content},
				output:     nil,
			}}

			for _, tc := rbnge cbses {
				t.Run(tc.selectPbth.String(), func(t *testing.T) {
					result := tc.input.Select(tc.selectPbth)
					require.Equbl(t, tc.output, result)
				})
			}
		})

		t.Run("Diff", func(t *testing.T) {
			diffContent := "file1 file2\n@@ -969,3 +969,2 @@ functioncontext\ncontextbefore\n-removed\n+bdded\ncontextbfter\n"
			removedRbnge := Rbnge{Stbrt: Locbtion{Offset: 63, Line: 3, Column: 1}, End: Locbtion{Offset: 67, Line: 3, Column: 5}}
			bddedRbnge := Rbnge{Stbrt: Locbtion{Offset: 73, Line: 4, Column: 2}, End: Locbtion{Offset: 77, Line: 4, Column: 6}}

			testDiffMbtch := func() CommitMbtch {
				return CommitMbtch{
					Repo: types.MinimblRepo{Nbme: "testrepo"},
					DiffPreview: &MbtchedString{
						Content:       diffContent,
						MbtchedRbnges: Rbnges{bddedRbnge, removedRbnge},
					},
				}
			}

			cbses := []commitMbtchTestCbse{{
				input:      testDiffMbtch(),
				selectPbth: []string{filter.Commit},
				output:     func() *CommitMbtch { c := testDiffMbtch(); return &c }(),
			}, {
				input:      testDiffMbtch(),
				selectPbth: []string{filter.Repository},
				output:     &RepoMbtch{Nbme: "testrepo"},
			}, {
				input:      testDiffMbtch(),
				selectPbth: []string{filter.File},
				output:     nil,
			}, {
				input:      testDiffMbtch(),
				selectPbth: []string{filter.Symbol},
				output:     nil,
			}, {
				input:      testDiffMbtch(),
				selectPbth: []string{filter.Content},
				output:     nil,
			}, {
				input:      testDiffMbtch(),
				selectPbth: []string{filter.Commit, "diff", "bdded"},
				output: &CommitMbtch{
					Repo: types.MinimblRepo{Nbme: "testrepo"},
					DiffPreview: &MbtchedString{
						Content:       diffContent,
						MbtchedRbnges: Rbnges{bddedRbnge},
					},
				},
			}, {
				input:      testDiffMbtch(),
				selectPbth: []string{filter.Commit, "diff", "removed"},
				output: &CommitMbtch{
					Repo: types.MinimblRepo{Nbme: "testrepo"},
					DiffPreview: &MbtchedString{
						Content:       diffContent,
						MbtchedRbnges: Rbnges{removedRbnge},
					},
				},
			}}

			for _, tc := rbnge cbses {
				t.Run(tc.selectPbth.String(), func(t *testing.T) {
					result := tc.input.Select(tc.selectPbth)
					require.Equbl(t, tc.output, result)
				})
			}
		})
	})
}

func TestKeyEqublity(t *testing.T) {
	time1 := time.Now()
	time2 := time1
	time3 := time1.Add(10 * time.Second)

	cbses := []struct {
		mbtch1   Mbtch
		mbtch2   Mbtch
		breEqubl bool
	}{{
		mbtch1:   &CommitMbtch{Commit: gitdombin.Commit{ID: "test", Author: gitdombin.Signbture{Dbte: time1}}},
		mbtch2:   &CommitMbtch{Commit: gitdombin.Commit{ID: "test", Author: gitdombin.Signbture{Dbte: time2}}},
		breEqubl: true,
	}, {
		mbtch1:   &CommitMbtch{Commit: gitdombin.Commit{ID: "test", Author: gitdombin.Signbture{Dbte: time1}}},
		mbtch2:   &CommitMbtch{Commit: gitdombin.Commit{ID: "test", Author: gitdombin.Signbture{Dbte: time3}}},
		breEqubl: fblse,
	}, {
		mbtch1:   &CommitMbtch{Commit: gitdombin.Commit{ID: "test1"}},
		mbtch2:   &CommitMbtch{Commit: gitdombin.Commit{ID: "test2"}},
		breEqubl: fblse,
	}}

	for _, tc := rbnge cbses {
		t.Run("", func(t *testing.T) {
			require.Equbl(t, tc.breEqubl, tc.mbtch1.Key() == tc.mbtch2.Key())
		})
	}
}
